+++ bugzoo/client/file.py
@@ -21,6 +21,30 @@ class FileManager(object):
         self.__api = api
         self.__mgr_bug = mgr_bug
 
+    def resolve(self, container: Container, fn: str) -> str:
+        """
+        Ensures that relative paths are transformed into absolute paths.
+        """
+        bug = self.__mgr_bug[container.bug]
+        fn_orig = fn
+
+        if not os.path.isabs(fn):
+            fn = os.path.join(bug.source_dir, fn)
+            logger.debug("converted relative path to absolute path: %s -> %s",
+                         fn_orig, fn)
+
+        return fn
+
+    def _file_path(self, container: Container, fn: str) -> str:
+        """
+        Computes the base path for a given file.
+        """
+        fn = self.resolve(container, fn)
+        assert fn[0] == '/'
+        fn = fn[1:]
+        path = "files/{}/{}".format(container.uid, fn)
+        return path
+
     def write(self,
               container: Container,
               filepath: str,
@@ -41,7 +65,7 @@ class FileManager(object):
         """
         logger.debug("writing to file [%s] in container [%s].",
                      filepath, container.uid)
-        path = "files/{}/{}".format(container.uid, filepath)
+        path = self._file_path(container, filepath)
         try:
             response = self.__api.put(path, data=contents)
             if response.status_code != 204:
@@ -72,7 +96,7 @@ class FileManager(object):
         """
         logger.debug("reading contents of file [%s] in container [%s].",
                      filepath, container.uid)
-        path = "files/{}/{}".format(container.uid, filepath)
+        path = self._file_path(container, filepath)
         response = self.__api.get(path)
         if response.status_code == 200:
             contents = response.text